[WEBDEV-606] Refactor SearchIndexPageSerializer - Split SearchIndexPageSerializer into SearchPage::LandingPageSerializer and SearchPage::ResultsPageSerializer - Update fixture helper methods to improve file path inference
