--- fixtures/spec/serializers/base_serializer/with_component_variant_modifier.yml
@@ -1,4 +1,0 @@
----
-:component: component
-:variant: variant
-:modifier: modifier
