--- fixtures/spec/serializers/base_serializer/with_content_and_variant.yml
@@ -1,3 +1,0 @@
----
-:component: component
-:variant: variant
