--- fixtures/spec/serializers/base_serializer/with_only_component.yml
@@ -1,2 +1,0 @@
----
-:component: component
