--- fixtures/spec/serializers/base_serializer/with_only_variant.yml
@@ -1,2 +1,0 @@
----
-:variant: variant
