+++ spec/controllers/search_controller_spec.rb
@@ -5,7 +5,7 @@ RSpec.describe SearchController, vcr: true do
   describe 'GET index' do
     context 'with no query' do
       before(:each) do
-        allow(PageSerializer::SearchIndexPageSerializer).to receive(:new)
+        allow(PageSerializer::SearchPage::LandingPageSerializer).to receive(:new)
 
         get :index
       end
@@ -15,7 +15,7 @@ RSpec.describe SearchController, vcr: true do
       end
 
       it 'calls the serializer correctly' do
-        expect(PageSerializer::SearchIndexPageSerializer).to have_received(:new)
+        expect(PageSerializer::SearchPage::LandingPageSerializer).to have_received(:new)
       end
     end
 
@@ -24,7 +24,7 @@ RSpec.describe SearchController, vcr: true do
         before(:each) do
           allow(controller.request).to receive(:env).and_return({'ApplicationInsights.request.id' => '|1234abcd.'})
 
-          allow(PageSerializer::SearchIndexPageSerializer).to receive(:new)
+          allow(PageSerializer::SearchPage::ResultsPageSerializer).to receive(:new)
         end
 
         before(:each) do
@@ -55,13 +55,13 @@ RSpec.describe SearchController, vcr: true do
         end
 
         it 'calls the serializer with the correct arguments' do
-          expect(PageSerializer::SearchIndexPageSerializer).to have_received(:new).with(opensearch_description_url: 'http://:/search/opensearch', query: 'banana', results: [1, 2, 3], pagination_hash: 'pagination_hash')
+          expect(PageSerializer::SearchPage::ResultsPageSerializer).to have_received(:new).with(opensearch_description_url: 'http://:/search/opensearch', query: 'banana', results: [1, 2, 3], pagination_hash: 'pagination_hash')
         end
       end
 
       context 'an invalid search' do
         before(:each) do
-          allow(PageSerializer::SearchIndexPageSerializer).to receive(:new)
+          allow(PageSerializer::SearchPage::ResultsPageSerializer).to receive(:new)
 
           get :index, params: { q: 'fdsfsd' }
         end
@@ -79,13 +79,13 @@ RSpec.describe SearchController, vcr: true do
               query: 'fdsfsd'
           }
 
-          expect(PageSerializer::SearchIndexPageSerializer).to have_received(:new).with(opensearch_description_url: 'http://test.host/search/opensearch', query: 'fdsfsd', results: an_instance_of(Feedjira::Parser::Atom), pagination_hash: pagination_hash)
+          expect(PageSerializer::SearchPage::ResultsPageSerializer).to have_received(:new).with(opensearch_description_url: 'http://test.host/search/opensearch', query: 'fdsfsd', results: an_instance_of(Feedjira::Parser::Atom), pagination_hash: pagination_hash)
         end
       end
 
       context 'with an empty string' do
         before(:each) do
-          allow(PageSerializer::SearchIndexPageSerializer).to receive(:new)
+          allow(PageSerializer::SearchPage::LandingPageSerializer).to receive(:new)
 
           get :index, params: { q: '' }
         end
@@ -95,7 +95,7 @@ RSpec.describe SearchController, vcr: true do
         end
 
         it 'calls the serializer with the correct arguments' do
-          expect(PageSerializer::SearchIndexPageSerializer).to have_received(:new).with(opensearch_description_url: 'http://test.host/search/opensearch', flash_message: I18n.t('search_controller.index.flash'))
+          expect(PageSerializer::SearchPage::LandingPageSerializer).to have_received(:new).with(opensearch_description_url: 'http://test.host/search/opensearch', flash_message: I18n.t('search_controller.index.flash'))
         end
       end
 
@@ -152,11 +152,11 @@ RSpec.describe SearchController, vcr: true do
 
           allow(Parliament::Request::OpenSearchRequest).to receive(:new) { open_search_request }
 
-          allow(PageSerializer::SearchIndexPageSerializer).to receive(:new) { double('serializer', to_h: true) }
+          allow(PageSerializer::SearchPage::ResultsPageSerializer).to receive(:new) { double('serializer', to_h: true) }
 
           get :index, params: { q: 'Allan Wazacz' }
 
-          expect(PageSerializer::SearchIndexPageSerializer).to have_received(:new).with(opensearch_description_url: 'http://test.host/search/opensearch', query: 'Allan Wazacz')
+          expect(PageSerializer::SearchPage::ResultsPageSerializer).to have_received(:new).with(opensearch_description_url: 'http://test.host/search/opensearch', query: 'Allan Wazacz')
         end
       end
     end
