+++ spec/fixtures/serializers/base_serializer/with_content_and_variant.yml
