+++ spec/fixtures/serializers/base_serializer/with_only_component.yml
