+++ spec/fixtures/serializers/base_serializer/with_only_variant.yml
