+++ spec/fixtures/serializers/page_serializer/search_page/landing_page_serializer/with_flash_message.yml
