+++ spec/fixtures/serializers/page_serializer/search_page/landing_page_serializer/without_query.yml
