+++ spec/fixtures/serializers/page_serializer/search_page/results_page_serializer/with_a_query.yml
